refactor(appserver): isolate thread handlers
